cleanup in gateways
